second round andrews comments wrt implem
