+++ deepinv/physics/__init__.py
@@ -46,7 +46,7 @@ from .noise import (
 from .mri import MRI, DynamicMRI, SequentialMRI, MultiCoilMRI
 from .wrappers import PhysicsMultiScaler, LinearPhysicsMultiScaler, PhysicsCropper
 from .tomography import Tomography, TomographyWithAstra
-from .functional.subsets import (
+from .functional.tomography_subsets import (
     get_subset_indices,
     split_measurements,
     split_physics,
