+++ docs/source/api/deepinv.physics.rst
@@ -112,7 +112,7 @@ Generators
    deepinv.physics.phase_retrieval.build_probe
    deepinv.physics.phase_retrieval.generate_shifts
    deepinv.physics.scattering.circular_sensors
-   deepinv.physics.functional.subsets.get_subset_indices
+   deepinv.physics.functional.tomography_subsets.get_subset_indices
    deepinv.physics.split_measurements
    deepinv.physics.split_physics
